requirements.txt
libusb
pyusb
